// services/statuspage_integration/src/v2/models/StatuspageUrlSettingUpdateAttributes.ts
import { AttributeTypeMap } from "@datadog/datadog-api-client";

/**
 * The Statuspage URL setting attributes for an update request.
 */
export class StatuspageUrlSettingUpdateAttributes {
  /**
   * Comma-separated list of custom tags to apply to events generated from this Statuspage URL.
   */
  "customTags"?: string;
  /**
   * The Statuspage URL to monitor.
   */
  "url"?: string;
  /**
   * A container for additional, undeclared properties.
   * This is a holder for any undeclared properties as specified with
   * the 'additionalProperties' keyword in the OAS document.
   */
  "additionalProperties"?: { [key: string]: any };
  /**
   * @ignore
   */
  "_unparsed"?: boolean;

  /**
   * @ignore
   */
  static readonly attributeTypeMap: AttributeTypeMap = {
    customTags: {
      baseName: "custom_tags",
      type: "string",
    },
    url: {
      baseName: "url",
      type: "string",
    },
    additionalProperties: {
      baseName: "additionalProperties",
      type: "{ [key: string]: any; }",
    },
  };

  /**
   * @ignore
   */
  static getAttributeTypeMap(): AttributeTypeMap {
    return StatuspageUrlSettingUpdateAttributes.attributeTypeMap;
  }

  public constructor() {}
}